// generated/cms-schemas.ts
/* eslint-disable */
/* prettier-ignore */
/* biome-ignore format: auto-generated */
// This file is auto-generated by cms-renderer. Do not edit manually.

import { z } from 'zod';

export const airportSchema = z.object({
  code: z.string().optional(),
  name: z.string().optional(),
  city: z.string().optional(),
  country: z.string().optional(),
  latitude: z.number().optional(),
  longitude: z.number().optional(),
});

export type Airport = z.infer<typeof airportSchema>;


export const bodySchema = z.object({
  code: z.string().optional(),
  city: z.string().optional(),
  country: z.string().optional(),
  latitude: z.number().optional(),
  longitude: z.number().optional(),
});

export type Body = z.infer<typeof bodySchema>;


export const footerSchema = z.object({
  text: z.string().optional(),
});

export type Footer = z.infer<typeof footerSchema>;


export const headlineSchema = z.object({
  title: z.string().optional(),
  subtitle: z.string().optional(),
});

export type Headline = z.infer<typeof headlineSchema>;


export const navSchema = z.object({
  brand: z.string().optional(),
});

export type Nav = z.infer<typeof navSchema>;